requirements.txt
flask
flask-cors
langchain
langchain-core
python-dotenv
openai 
langchain-openai